chore: Extract models
